hecho lo relacionado con espacios publicos (crear, actualizar, borrar). Falla un error tonto

+++ src/main/java/com/ams/ei1027espaciosnaturales/controller/EspacioPublicoController.java
@@ -4,6 +4,7 @@ package com.ams.ei1027espaciosnaturales.controller;
 import com.ams.ei1027espaciosnaturales.dao.EspacioPublicoDAO;
 import com.ams.ei1027espaciosnaturales.model.EspacioPublico;
 
+import com.ams.ei1027espaciosnaturales.model.UserInterno;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DuplicateKeyException;
@@ -15,11 +16,16 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import javax.servlet.http.HttpSession;
+import java.util.Arrays;
+import java.util.List;
+
 @Controller
 @RequestMapping("/espacioPublico")
 public class EspacioPublicoController {
 
     private EspacioPublicoDAO espacioPublicoDAO;
+    private static final EspacioPublicoValidator validator = new EspacioPublicoValidator();
 
     @Autowired
     public void setEspacioPublicoDAO(EspacioPublicoDAO e) {
@@ -35,7 +41,13 @@ public class EspacioPublicoController {
 
     // Los siguientes dos metodos gestionan la inserción de un espacio publico
     @RequestMapping(value = "/add")
-    public String addEspacioPublico(Model model) {
+    public String addEspacioPublico(HttpSession session, Model model) {
+        UserInterno user = checkSession(session);
+        if (user == null){
+            model.addAttribute("user", new UserInterno());
+            return "login";
+        }
+
         model.addAttribute("espacioPublico", new EspacioPublico());
         return "espacioPublico/add";
     }
@@ -43,6 +55,7 @@ public class EspacioPublicoController {
     @RequestMapping(value = "/add", method = RequestMethod.POST)
     public String processAddEspacioPublico(@ModelAttribute("espacioPublico") EspacioPublico e,
                                       BindingResult bindingResult) {
+        validator.validate(e, bindingResult);
         if (bindingResult.hasErrors()) {
             return "espacioPublico/add";
         }
@@ -53,16 +66,28 @@ public class EspacioPublicoController {
             throw new EspaciosNaturalesException("Ya existe un espacio público con el mismo nombre",
                     "CPDuplicada", "espacioPublico/add");
         }
-        catch (DataAccessException exception){
-            throw new EspaciosNaturalesException("Error accediendo a la base de datos", "ErrorAccidiendoDatos", "/");
-        }
+//        catch (DataAccessException exception){
+//            throw new EspaciosNaturalesException("Error accediendo a la base de datos", "ErrorAccidiendoDatos", "/");
+//        }
         return "redirect:list";
     }
 
     // Los siguientes dos metodos gestionan la modificacion de un espacio publico
     @RequestMapping(value = "/update/{nombre}", method = RequestMethod.GET)
-    public String updateEspacioPublico(Model model, @PathVariable String nombre) {
+    public String updateEspacioPublico(HttpSession session, Model model, @PathVariable String nombre) {
+        UserInterno user = checkSession(session);
+        if (user == null){
+            model.addAttribute("user", new UserInterno());
+            return "login";
+        }
+
         model.addAttribute("espacioPublico", espacioPublicoDAO.getEspacioPublico(nombre));
+        List<String> espacios = Arrays.asList("playa", "río", "estanque", "lago", "bosque", "otros");
+        model.addAttribute("espacioList", espacios);
+        List<String> suelos = Arrays.asList("arena", "roca", "piedra");
+        model.addAttribute("sueloList", suelos);
+        List<String> accesos = Arrays.asList("abierto", "restringido", "cerrado");
+        model.addAttribute("accesoList", accesos);
         return "espacioPublico/update";
     }
 
@@ -75,8 +100,28 @@ public class EspacioPublicoController {
     }
 
     @RequestMapping(value = "/delete/{nombre}")
-    public String processDeleteEspacioPublico(@PathVariable String nombre) {
+    public String processDeleteEspacioPublico(HttpSession session, Model model, @PathVariable String nombre) {
+        UserInterno user = checkSession(session);
+        if (user == null){
+            model.addAttribute("user", new UserInterno());
+            return "login";
+        }
+
         espacioPublicoDAO.deleteEspacioPublicoNombre(nombre);
         return "redirect:../list";
+    }
+
+    private UserInterno checkSession(HttpSession session){
+        if(session.getAttribute("user") == null) return null;
+
+        UserInterno user = (UserInterno) session.getAttribute("user");
+
+        if (!user.getRol().equals("gestor")) {
+            System.out.println("El usuario no puede acceder a esta pagina con este rol");
+            throw new EspaciosNaturalesException("No tienes permisos para acceder a esta página porque no eres un gestor",
+                    "AccesDenied", "../" + user.getUrlMainPage());
+        }
+
+        return user;
     }
 }
